Refactor AbsTask
move AbsTask.run() -> Trainer.run()
move AbsTask.average_nbest_models() -> average_nbest_models()
move AbsTask.collect_stats() -> collect_stats()
move DistributedOption -> espnet2/train/distributed_utils.py
